Moving the entities to a different folder for project readability
