+++ MontyHallKata/Models/Entity/Door.cs
@@ -1,10 +1,9 @@
-namespace MontyHallKata.Models
+namespace MontyHallKata.Models.Entity
 {
     public class Door
     {
         public bool IsOpen { get; set; }
         public bool IsSelected { get; set; }
-        
         public bool IsWinningDoor { get; init; }
     }
 }
