+++ MontyHallTests/SimulationTests.cs
@@ -2,6 +2,7 @@ using System;
 using MontyHallKata;
 using MontyHallKata.Controllers;
 using MontyHallKata.Models;
+using MontyHallKata.Models.Entity;
 using MontyHallKata.Models.Randomizer;
 using Moq;
 using Xunit;
